When a line is added to a Skirmish/Fight, raise a property changed for the whole object.
While this is a brute force approach to detecting exactly which property changed within a Fight, it is reasonable. A new line will end up affecting multiple properties down the whole property hierarchy. It is easier to just figure that the whole thing has changed. The ViewModel that will update itself for these property changes will have a more precise property change for any UI properties that actually change.