Remove extra entries in the name split.
We're looking for lower case parts of the name, an empty entry ("") is not necessary to check.
An empty entry can occur if there is multiple spaces in a name.
This occurs in this case, because the WhoParse wasn't parsing the name correctly and there are extra spaces if a character is "  LFG" (note two space)